# api-documentation-old.yaml
openapi: 3.0.0
info:
  title: InCrowd API
  version: 1.0.0
  description: |
    # InCrowd API - Documentazione Completa
    
    Benvenuto nella documentazione dell'API di InCrowd, una piattaforma innovativa di crowdsourcing per proposte e iniziative cittadine.
    
    ## Panoramica
    
    L'API InCrowd fornisce accesso completo alla piattaforma per:
    - Gestione utenti (cittadini privati ed enti)
    - Creazione e gestione proposte
    - Sistema di follow/followers
    - Amministrazione e moderazione
    
    ## Autenticazione e Sicurezza
    
    L'API utilizza autenticazione JWT (JSON Web Token):
    
    1. **Login**: Ottieni il token tramite `POST /api/auth/login`
    2. **Utilizzo**: Includi il token nell'header di ogni richiesta:
       ```
       Authorization: Bearer <your-jwt-token>
       ```
    3. **Scadenza**: I token hanno durata limitata, effettua refresh quando necessario
    
    ## Struttura delle Response
    
    Tutte le risposte seguono il formato standardizzato:
    
    ```json
    {
      "data": { /* Dati della risposta */ },
      "message": "Messaggio descrittivo",
      "error": null // Solo presente in caso di errori
    }
    ```
    
    ## Gestione Errori
    
    - **400 Bad Request**: Dati non validi o parametri mancanti
    - **401 Unauthorized**: Token mancante, scaduto o non valido
    - **403 Forbidden**: Permessi insufficienti per l'operazione
    - **404 Not Found**: Risorsa non trovata
...
    
    ## Tipi di Utente
    
...
    
...
    
...

...
servers:
...
paths:
...
  /api/operatori:
    post:
      tags: [Operatori]
      operationId: createOperatore
      summary: Crea un nuovo operatore (solo admin)
      description: |
        Crea un nuovo account operatore nel sistema.
        Disponibile solo per amministratori.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nome
                - cognome
                - email
                - password
              properties:
                nome:
                  type: string
                  example: "Mario"
                cognome:
                  type: string
                  example: "Rossi"
                email:
                  type: string
                  format: email
                  example: "mario.operatore@incrowd.com"
                password:
                  type: string
                  minLength: 8
                  example: "SecurePass123!"
      responses:
        '201':
          description: Operatore creato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
              example:
                data:
                  operatore:
                    _id: "60d0fe4f5311236168a109cf"
                    nome: "Mario"
                    cognome: "Rossi"
                    email: "mario.operatore@incrowd.com"
                    createdAt: "2024-08-18T10:30:00Z"
                message: "Operator created successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/operatori/admin:
    get:
      tags: [Operatori]
      operationId: getAllOperatori
      summary: Recupera tutti gli operatori (solo admin)
      description: |
        Ottiene la lista di tutti gli operatori registrati nel sistema.
        Endpoint riservato esclusivamente agli amministratori.
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - in: query
          name: search
          schema:
            type: string
          description: "Termine di ricerca per nome o cognome"
      responses:
        '200':
          description: Lista operatori recuperata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
              example:
                data:
                  operatori:
                    - _id: "60d0fe4f5311236168a109cf"
                      nome: "Mario"
                      cognome: "Rossi"
                      email: "mario.operatore@incrowd.com"
                      createdAt: "2024-01-15T10:30:00Z"
                      stats:
                        totalModerated: 245
                        approved: 198
                        rejected: 47
                    - _id: "60d0fe4f5311236168a109d0"
                      nome: "Luigi"
                      cognome: "Verdi"
                      email: "luigi.operatore@incrowd.com"
                      createdAt: "2024-02-20T14:15:00Z"
                      stats:
                        totalModerated: 189
                        approved: 156
                        rejected: 33
                  pagination:
                    currentPage: 1
                    totalPages: 2
                    totalItems: 25
                    itemsPerPage: 20
                message: "Operators retrieved successfully"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Operatori]
      operationId: createOperatoreAdmin
      summary: Crea nuovo operatore (solo admin)
      description: |
        Crea un nuovo account operatore nel sistema.
        Endpoint riservato esclusivamente agli amministratori.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nome
                - cognome
                - email
                - password
              properties:
                nome:
                  type: string
                  example: "Mario"
                cognome:
                  type: string
                  example: "Rossi"
                email:
                  type: string
                  format: email
                  example: "mario.operatore@incrowd.com"
                password:
                  type: string
                  minLength: 8
                  example: "SecurePass123!"
      responses:
        '201':
          description: Operatore creato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
              example:
                data:
                  operatore:
                    _id: "60d0fe4f5311236168a109cf"
                    nome: "Mario"
                    cognome: "Rossi"
                    email: "mario.operatore@incrowd.com"
                    createdAt: "2024-08-18T10:30:00Z"
                message: "Operator created successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/operatori/admin/{id}:
    delete:
      tags: [Operatori]
      operationId: deleteOperatore
      summary: Elimina operatore (solo admin)
      description: |
        Elimina un operatore dal sistema. Disponibile solo per amministratori.
        L'operazione è irreversibile.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: "60d0fe4f5311236168a109ca"
          description: "ID dell'operatore da eliminare"
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Operatore eliminato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
              example:
                message: "Operator deleted successfully"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT token ottenuto tramite login"
  responses:
    BadRequest:
      description: "Richiesta non valida - dati mancanti o formato errato"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardResponse'
          example:
            message: "Invalid request data"
            error:
              code: "BAD_REQUEST"
              details: "The provided data is not valid"
    Unauthorized:
      description: "Non autorizzato - token mancante o non valido"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardResponse'
          example:
            message: "Authentication required"
            error:
              code: "UNAUTHORIZED"
              details: "Token is missing or invalid"
    Forbidden:
      description: "Accesso negato - privilegi insufficienti"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardResponse'
          example:
            message: "Access denied"
            error:
              code: "FORBIDDEN"
              details: "Insufficient privileges for this operation"
    NotFound:
      description: "Risorsa non trovata"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardResponse'
          example:
            message: "Resource not found"
            error:
              code: "NOT_FOUND"
              details: "The requested resource does not exist"
    Conflict:
      description: "Conflitto - risorsa già esistente"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardResponse'
          example:
            message: "Resource already exists"
            error:
              code: "CONFLICT"
              details: "Email address already registered"
    InternalError:
      description: "Errore interno del server"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardResponse'
          example:
            message: "Internal server error"
            error:
              code: "INTERNAL_SERVER_ERROR"
              details: "An unexpected error occurred"
  schemas:
...
    StandardResponse:
      type: object
      properties:
        data:
          type: object
          description: "Contenuto della risposta (opzionale in caso di errore)"
        message:
          type: string
          description: "Messaggio informativo sempre presente"
          example: "Operation completed successfully"
        error:
          type: object
          description: "Dettagli errore (solo in caso di errore)"
          properties:
            code:
              type: string
              example: "ERROR_CODE"
            details:
              type: string
              example: "Detailed error description"
      required:
        - message

    # =============================================================================
    # USER SCHEMA (UNIFICATO)
    # =============================================================================
...
security:
  - BearerAuth: []
  - {}